Add pump safer
As pump may stand still for quiet some time, it may break due lack of
motion. To prevent this, the pump will now be turned on once within 24h.
Additionally the time before switching the pump on or off again is
increased to 30 minutes.

this should avoid the required patching
of the `Adafruit_SSD1306.h`